Added first Statements

+++ Tools/GenerateAst/generate_ast.py
@@ -2,9 +2,10 @@ import os
 import sys
 
 def define_ast(output_dir, basename, types):
-    path = f"{output_dir}/{basename}.bf"
+    path = f"{output_dir}\\{basename}.bf"
 
-    os.remove(path)
+    if os.path.exists(path):
+        os.remove(path)
 
     with open(path, 'w') as file:
         file.writelines([
@@ -13,12 +14,15 @@ def define_ast(output_dir, basename, types):
              "{\n",
             f"    public abstract class {basename}\n",
              "    {\n",
-             "        // Virtual/Abstract generic are not yet supported\n"
-             "        public abstract void Accept(Visitor visitor);\n",
-             "    }\n\n"
         ])
 
         define_visitor(file, basename, types)
+
+        file.writelines([
+            "        // Virtual/Abstract generic are not yet supported\n"
+            "        public abstract void Accept(Visitor visitor);\n",
+            "    }\n\n"
+        ])
 
         for str in types:
             classname = str.split(':')[0].strip()
@@ -31,20 +35,20 @@ def define_ast(output_dir, basename, types):
 
 def define_visitor(file, basename, types):
     file.writelines([
-        "    public interface Visitor\n",
-        "    {\n"
+        "        public interface Visitor\n",
+        "        {\n"
     ])
 
     for type in types:
         name = type.split(':')[0].strip()
-        file.write(f"        void Visit{name}{basename}({name} {basename.lower()});\n")
+        file.write(f"            void Visit{name}{basename}({name} {basename.lower()});\n")
 
-    file.write("    }\n\n")
+    file.write("        }\n\n")
     file.writelines([
-        "    public interface Visitor<T>: Visitor\n",
-        "    {\n"
-        "        public T Result { get; };\n"
-        "    }\n"
+        "        public interface Visitor<T>: Visitor\n",
+        "        {\n"
+        "            public T Result { get; };\n"
+        "        }\n"
     ])
 
 def define_types(file, basename, classname, fieldList):
@@ -91,15 +95,16 @@ if __name__ == "__main__":
         sys.exit(64)
 
     output_dir = sys.argv[1]
+    
     define_ast(output_dir, "Expr", [
-        #"Equality   : ",
-        #"Comparison : ",
-        #"Term       : ",
-        #"Factor     : ",
         "Unary		: Token op, Expr right",
-        #"Primary    : ",
         "Binary		: Expr left, Token op, Expr right",
         "Grouping	: Expr expression",
         "Literal	: Variant value",
+    ])
+
+    define_ast(output_dir, "Stmt", [
+        "Expression : Expr expression",
+        "Print      : Expr expression"
     ])
